requirements-fastapi.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Google Cloud
google-cloud-firestore==2.13.0
google-cloud-storage==2.10.0
google-cloud-logging==3.8.0
google-cloud-trace==1.11.0
google-cloud-monitoring==2.15.1
google-cloud-run==0.10.0

# AI/ML
langgraph==0.0.20
langchain==0.1.0
google-generativeai==0.3.0

# Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-gcp-trace==1.5.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
prometheus-client==0.19.0

# Security & Auth
python-jose==3.3.0
passlib==1.7.4
pyjwt==2.8.1
cryptography==41.0.7

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiohttp==3.9.1
pydantic-core==2.14.1
pytz==2023.3
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Logging
structlog==23.2.0
python-json-logger==2.0.7
